Now supports all compilers

+++ Core/Inc/battery_model.h
@@ -7,12 +7,12 @@
 
 typedef struct {
     
-    uint8_t CellID; // No battery in align until this point has had more than 255 cells
-    uint8_t TemperatureID; // What sensor to listen to for temperature
-    uint16_t EstimatedVoltage;
-    int16_t EstimatedCurrent;
-    uint16_t MeasuredVoltage;
-    int8_t MeasuredTemperature;
+    uint16_t CellID; // No battery in align until this point has had more than 255 cells
+    uint16_t TemperatureID; // What sensor to listen to for temperature
+    float EstimatedVoltage;
+    float EstimatedCurrent;
+    float MeasuredVoltage;
+    float MeasuredTemperature;
 
 } CellModel_HandleTypeDef;
 
@@ -26,8 +26,8 @@ typedef struct {
 
 } BatteryModel_HandleTypeDef;
 
-void BatteryModel_Init(BatteryModel_HandleTypeDef *battery_model, uint8_t cell_count);
-void BatteryModel_UpdateMeasured(BatteryModel_HandleTypeDef *battery_model, uint16_t *cell_voltages, int8_t *die_temperatures);
+void BatteryModel_Init(BatteryModel_HandleTypeDef *battery_model, uint16_t cell_count);
+void BatteryModel_UpdateMeasured(BatteryModel_HandleTypeDef *battery_model, float *cell_voltages, float *die_temperatures);
 
 
 #endif // BATTERY_MODEL_H
